refactor: everything in its right place
